feat: crud de lojas completa funcionando com autenticação

+++ backend/src/stores/stores.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/database/prisma.service';
 import { CreateStoreDto, UpdateStoreDto } from './stores.dto';
 
@@ -31,12 +31,17 @@ export class StoresService {
     return storeExists;
   }
 
-  async update(id: number, data: UpdateStoreDto) {
+  async update(id: number, data: UpdateStoreDto, userId: number) {
     const storeExists = await this.prisma.stores.findUnique({ 
     where: {id} 
   })
   if(!storeExists){
     throw new NotFoundException("Loja não encontrada")
+  }
+  if (storeExists.user_id !== userId){
+    throw new ForbiddenException(
+      'Voce não tem permissão para editar esta loja'
+    );
   }
   return this.prisma.stores.update({
       where: {id},
@@ -44,12 +49,17 @@ export class StoresService {
     });
   }
 
-  async remove(id: number) {
+  async remove(id: number, userId: number) {
     const storeExists = await this.prisma.stores.findUnique({ 
     where: {id} 
   })
   if(!storeExists){
     throw new NotFoundException("Loja não encontrada")
+  }
+  if(storeExists.user_id !== userId){
+    throw new ForbiddenException(
+      'Você não tem permissão para deletar essa loja'
+    );
   }
     return this.prisma.stores.delete({ where: { id } });
   }
